fix: review fixes — type hints, broken links, provider placeholders, patch format, and placeholder consistency
- Fix _CacheKey type hint (Tuple[str, float] -> Tuple[str, int])
- Fix respond.yml relative paths (agent/...) -> full GitHub URLs
- Fix env_loader.py.patch: replaced stub with complete copy-paste instructions
- Fix test-script.sh.example: replaced system-specific redaction format
  with standard sk-placeholder values
- Fix README encrypted backup section: replaced hardcoded 'opencode-go'
  provider name with PROVIDER variable and generic credentials.keys.enc path
- Fix security checklist: consistent path notation (auth.json instead
  of /root/.hermes/auth.json, .env instead of ~/.hermes/.env)

+++ agent/secret_sources/script.py
@@ -59,7 +59,7 @@ logger = logging.getLogger(__name__)
 # In-process cache so back-to-back load_hermes_dotenv() calls don't
 # re-execute the script on every call.
 # ---------------------------------------------------------------------------
-_CacheKey = Tuple[str, float]  # (command, config_fingerprint)
+_CacheKey = Tuple[str, int]  # (command, hash_of_config)
 _CACHE: Dict[_CacheKey, "_CachedFetch"] = {}
